Extract shared mesh LLM routing crate
