+++ Cargo.toml
@@ -3,6 +3,7 @@ members = [
     "crates/mesh-llm",
     "crates/mesh-llm-identity",
     "crates/mesh-llm-protocol",
+    "crates/mesh-llm-routing",
     "crates/mesh-llm-types",
     "crates/mesh-llm-ui",
     "crates/mesh-llm-plugin",
