+++ crates/mesh-client/README.md
@@ -31,6 +31,10 @@ client call sites.
 Shared protobuf types and frame helpers are owned by `crates/mesh-llm-protocol/`
 and re-exported here for compatibility with existing client call sites.
 
+Shared routing targets and model placement helpers are owned by
+`crates/mesh-llm-routing/` and re-exported here for compatibility with existing
+client call sites. Client-only runtime process details stay in this crate.
+
 Client requests should preserve the full model ref chosen by the caller. Model
 resolution, stage topology, and runtime lifecycle remain server-side mesh
 responsibilities.
