+++ crates/mesh-llm/Cargo.toml
@@ -8,6 +8,7 @@ ansi-to-tui = "8"
 mesh-llm-plugin = { path = "../mesh-llm-plugin" }
 mesh-llm-identity = { path = "../mesh-llm-identity" }
 mesh-llm-protocol = { path = "../mesh-llm-protocol" }
+mesh-llm-routing = { path = "../mesh-llm-routing" }
 mesh-llm-types = { path = "../mesh-llm-types" }
 mesh-llm-ui = { path = "../mesh-llm-ui" }
 mesh-client = { package = "mesh-llm-client", path = "../mesh-client", features = ["host-io"] }
